requirements.txt
rpi-hardware-pwm==0.1.3
RPi.GPIO==0.7.0
paho-mqtt==1.6.1